Move VedeuCommonClass inline.
