+++ test/lib/vedeu/support/common_test.rb
@@ -2,6 +2,16 @@ require 'test_helper'
 
 module Vedeu
 
+  class VedeuCommonClass
+
+    include Vedeu::Common
+
+    def defined_value_test(variable)
+      defined_value?(variable)
+    end
+
+  end # VedeuCommonClass
+
   describe Common do
 
     let(:described) { Vedeu::VedeuCommonClass.new }
